refactor: Add replaceText method to InputBuffer and SpeechSynthesisModule
This commit adds a new method, replaceText, to the InputBuffer class and the SpeechSynthesisModule class. The replaceText method allows for replacing specific text in the buffer and speech stream, respectively. This functionality enhances the flexibility and customization of the text-to-speech feature by providing the ability to dynamically modify the speech content during runtime.